+++ src/rotationalparticle.cpp
@@ -61,12 +61,12 @@ bool RotationalParticle::update()
 
     float rad = static_cast<float>(atan2(mVelocity.x, mVelocity.y));
     if (rad < 0)
-        rad = static_cast<float>(PI + (PI + rad));
+        rad = static_cast<float>(PI) + static_cast<float>(PI) + rad;
 
     float range = static_cast<float>(PI / size);
 
     // Determines which frame the particle should play
-    if (rad < range || rad > ((PI*2) - range))
+    if (rad < range || rad > ((static_cast<float>(PI)*2) - range))
     {
         mAnimation->setFrame(0);
     }
